requirements.txt
absl-py==2.1.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
backcall==0.2.0
boto3==1.34.158
botocore==1.34.158
cachetools==5.4.0
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
dataclasses==0.6
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dlib==19.24.99
executing==2.0.1
facenet-pytorch==2.5.3
filelock==3.15.4
filetype==1.2.0
fire==0.6.0
fonttools==4.50.0
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.33.0
google-auth-oauthlib==1.0.0
grpcio==1.65.4
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
imutils==0.5.4
ipykernel==6.29.3
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lapx==0.5.9.post1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mish-cuda==0.0.3
mmcv-full==1.4.8
-e git+https://github.com/zgchen33/MCGaze.git@61950c878dcb4f58675863dcbf0693b5311d2743#egg=mmdet
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
packaging==24.0
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==5.27.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybboxes==0.1.6
pycocotools==2.0.7
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
roboflow==1.1.37
rsa==4.9
s3transfer==0.10.2
sahi==0.11.18
scikit-learn==1.3.2
scipy==1.10.0
seaborn==0.13.2
shapely==2.0.5
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
supervision==0.22.0
sympy==1.13.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tomli==2.0.1
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tornado==6.4
tqdm==4.19.9
traitlets==5.14.2
triton==3.0.0
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.0.152
ultralytics-thop==2.0.0
urllib3==1.26.19
wcwidth==0.2.13
Werkzeug==3.0.3
yapf==0.40.1
yolov5==7.0.13
zipp==3.18.1
